% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_correct_synoptics.R
\name{assert_correct_synoptics}
\alias{assert_correct_synoptics}
\title{Check correct synoptics format}
\usage{
assert_correct_synoptics(x)
}
\arguments{
\item{object}{that needs to be assessed}
}
\value{
stop message or data.frame
}
\description{
Check correct synoptics format
}
\examples{
test <- data.frame(
  syntaxonCode = rep("000", 2),
  speciesNumber = c(437, 463),
  frequency = c(1.1, 3.9),
  mean_if_present = c(2.43, 2.21)
)
assert_correct_synoptics(test)

}
