% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_indices.R
\name{classify_cod}
\alias{classify_cod}
\title{Calculate Composite Distance (CoD) (Eq. 9)}
\usage{
classify_cod(analysis_df, r = 1.5, s = 0.6, ...)
}
\arguments{
\item{analysis_df}{A data frame prepared by \code{create_analysis_data()}.}

\item{r}{Weight for Weirdness component (default 1.5).}

\item{s}{Power for MED component (default 0.60).}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class \code{inbovegclassification_list}.
}
\description{
Internal function to calculate the Composite Distance.
}
\keyword{internal}
