% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_incompleteness.R,
%   R/classify_incompleteness_old.R
\name{classify_incompleteness}
\alias{classify_incompleteness}
\title{Classify based on Incompleteness Index}
\usage{
classify_incompleteness(data, synoptics)

classify_incompleteness(data, synoptics)
}
\arguments{
\item{data}{A dataframe of plot data.}

\item{synoptics}{A dataframe of synoptic data.}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class 'inbovegclassification_list'.

An inbovegclassification object
}
\description{
Implements Equation 6 from van Tongeren et al. (2008) \link[=cite: 1]{cite_start}.
Calculates the sum of contributions from species absent from the relevé.

Calculates incompleteness index following van Tongeren et al. (2008)
}
