% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_likelihood.R
\name{classify_likelihood}
\alias{classify_likelihood}
\title{Calculate Likelihood Index for Vegetation Classification}
\usage{
classify_likelihood(data, synoptics = "default", normalised = TRUE)
}
\arguments{
\item{data}{A dataframe containing the plot data with columns:
\itemize{
\item RecordingGivid: Unique identifier for the recording
\item LayerCode: Code identifying vegetation layer
\item CoverageCode: Code for coverage type
\item PctValue: Percentage cover value
\item gbif_usageKey: GBIF species identifier
}
on wich recordings are identified.
When NULL all data is assumed from same recording}

\item{synoptics}{A dataframe containing reference syntaxa data with columns:
\itemize{
\item syntaxoncode: Code identifying the vegetation type
\item usageKey: GBIF species identifier
\item frequentie: Frequency of species in vegetation type (0-1)
}
@param normalised ignored logical variable}
}
\value{
A dataframe with columns:
\itemize{
\item syntaxoncode: Vegetation type code
\item likelihood: -2ln(likelihood) value,
where lower values indicate better matches
}
}
\description{
Calculates the maximum likelihood index to classify vegetation plots
against reference syntaxa,
following the ASSOCIA method (van Tongeren et al. 2008).
This function processes one record at a time.
}
\details{
The function:
\enumerate{
\item Filters for herb layer ('K')
\item Converts percentage cover to fractions
\item Calculates likelihood components for present and absent species
\item Sums likelihood values by syntaxon
}
}
\references{
van Tongeren, O., Gremmen, N., & Hennekens, S. (2008). Assignment of relevés
to pre-defined classes by supervised clustering of plant communities
using a new composite index.
Journal of Vegetation Science, 19(4), 525-536.
}
