% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_indices.R
\name{classify_med}
\alias{classify_med}
\title{Calculate Modified Euclidean Distance (MED) (Eq. 8)}
\usage{
classify_med(analysis_df, t = 1, ...)
}
\arguments{
\item{analysis_df}{A data frame prepared by \code{create_analysis_data()}.}

\item{t}{Weight for absent species in MED (default 1, as per paper).}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class \code{inbovegclassification_list}.
}
\description{
Internal function to calculate the Modified Euclidean Distance.
}
\keyword{internal}
