% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db_flora.R
\name{connect_db_flora}
\alias{connect_db_flora}
\title{Connect to INBO flora database}
\usage{
connect_db_flora(db = "D0012_00_Flora", test = FALSE)
}
\arguments{
\item{db}{The name of the INBO hosted database. Defaults to Taxa db}

\item{test}{flag whether the test modus is used}
}
\value{
A database connection object

connection object
}
\description{
Creates a database connection to the INBO flora database.
This function is a wrapper around \code{connect_inbo_dbase()}
specifically for connecting to the taxonomy database.
}
\details{
This function establishes a connection to the INBO taxonomy database
which contains standardized
taxonomic information. The database contains authoritative taxonomic data
that can be used
for standardizing species names and checking taxonomic classifications.
}
\section{Database}{

The taxonomy database (D0155_00_Taxa) contains standardized
taxonomic information used by INBO.
}

