% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db_connect.R
\name{connect_db_inboveg}
\alias{connect_db_inboveg}
\title{Connect to the INBOVEG database}
\usage{
connect_db_inboveg(db = "D0010_00_Cydonia", test = FALSE)
}
\arguments{
\item{db}{The name of the INBO hosted database. Defaults to "D0010_00_Cydonia".}

\item{test}{Flag whether the test modus is used. If TRUE, tries to connect
to a local test database (requires setup).}
}
\value{
A database connection object.
}
\description{
Connect to the INBOVEG database
}
