% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_classification_data.R
\name{create_classification_data}
\alias{create_classification_data}
\title{Create a dataset ready for classification}
\usage{
create_classification_data(data, synoptics)
}
\arguments{
\item{data}{data with recordings}

\item{synoptics}{synoptic table when data.frame or label to get the table}
}
\value{
combined data.frame with all necessary info for classification
}
\description{
Create a dataset ready for classification
}
