% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_inboveg_database.R
\name{create_test_inboveg_database}
\alias{create_test_inboveg_database}
\title{Create a test database}
\usage{
create_test_inboveg_database()
}
\value{
inboveg database
}
\description{
Create a test database
}
