% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_synoptics}
\alias{example_synoptics}
\title{Example synoptic classification table}
\format{
A data frame with 7800 rows and 7 variables (24 syntaxons):
\describe{
\item{syntaxonCode}{Recording identifier}
\item{speciesNumber}{Species identifier}
\item{frequency}{Presence frequency of a species in a syntaxon}
\item{mean_if_present}{The average frequency when present in a syntaxon}
\item{row_id}{Rownumber for use when there are several GBIF usageKeys for
the same species }
\item{usageKey}{GBIF usageKey of the species}
\item{acceptedUsageKey}{GBIF accepted usageKey when a synonym}
}
}
\source{
Subset of syntaxons that can be classified based on vegetation
}
\usage{
example_synoptics
}
\description{
A dataset containing the necessary info for classification of vegetation
in syntaxons.
}
\keyword{datasets}
