% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_getters.R
\name{get_synoptic_table}
\alias{get_synoptic_table}
\title{Get the synoptic table}
\usage{
get_synoptic_table(synoptics)
}
\arguments{
\item{synoptics}{A data frame with the synoptic table, the string
"default" to use the package's default table, or a file path
to a .csv2 file.}
}
\value{
A data frame containing the synoptic table, validated to have
the correct columns and data types.
}
\description{
Loads the synoptic table from the package's internal data or a user-provided
path.
}
