% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_flora_db.R
\name{link_flora_db}
\alias{link_flora_db}
\title{Link species records to GBIF taxonomy via INBO flora database}
\usage{
link_flora_db(con, species_names)
}
\arguments{
\item{con}{A database connection object to the INBO taxonomy database}

\item{species_names}{A character vector containing species records}
}
\value{
A dataframe containing the matched species with the futondb
}
\description{
This function links species records to their corresponding
florabank taxonomy information
by matching species names against the INBO taxonomy database (FUTON source).
It performs a partial string match to accommodate variations
in species name notation.
}
\examples{
\dontrun{
library(inbovegtypering)
con <- connect_db_flora()
species_names <- c("Dryopteris", "Nardus stricta L.",
                   "Lophocolea", "Fraxinus excelsior L.")
link_futon_db(con, species_names = species_names)
}

}
\seealso{
\code{\link{connect_db_flora}} for establishing the database connection
}
