% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_synoptic_species.R
\name{link_synoptic_species}
\alias{link_synoptic_species}
\title{Try to find the corresponding species_number in de synoptic table}
\usage{
link_synoptic_species(x, which = "name")
}
\arguments{
\item{x}{record for which the synoptic species number must be found.
Max 1 row. Containing at least columns usageKey and Originalname}

\item{which}{character controlling what is to be tested
possibilities are "usage_key", "genus_key", "name"}
}
\value{
data.frame with column speciesNumber
}
\description{
Try to find the corresponding species_number in de synoptic table
}
\examples{
library(inbovegtypering)
test <- data.frame(OriginalName = "Lemna trisulca", usageKey = 2867579)
link_synoptic_species(test$usageKey, which = "usage_key")
link_synoptic_species(test$OriginalName, which = "name")

test2 <- data.frame(name = c("Potamogeton pectinatus","Algenvlokken",
                             "Carex", "Carex Nonexistans"))
test2 |>
rowwise() |>
mutate(speciesNumber = link_synoptic_species(name, which = "name"))

}
