% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_species_list.R
\name{load_species_list}
\alias{load_species_list}
\title{Load species list}
\usage{
load_species_list(source = "package")
}
\arguments{
\item{source}{source where to find the species list}
}
\value{
tibble with important species fields
}
\description{
Load species list
}
