% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_synoptic_data.R
\name{load_synoptic_data}
\alias{load_synoptic_data}
\title{Load synoptic data}
\usage{
load_synoptic_data(source = "test")
}
\arguments{
\item{source}{location where to find the data}
}
\value{
tibble with synoptic data
}
\description{
Synoptic data are used as basis for the classification
}
