% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods_plot.R
\name{plot.inbovegclassification}
\alias{plot.inbovegclassification}
\title{Plot Method for inbovegclassification Objects}
\usage{
\method{plot}{inbovegclassification}(x, types = 12, ...)
}
\arguments{
\item{x}{An object of class "inbovegclassification"}

\item{types}{Either a number indicating how many top types to show
(default 12), or a character vector of syntaxon codes to display.}

\item{...}{Additional arguments passed to \code{plot_classification_rose()}}
}
\value{
A ggplot object
}
\description{
Generic plot method that dispatches to \code{plot_classification_rose()}.
}
