% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inbovegclassification.R
\name{plot.inbovegclassification_list}
\alias{plot.inbovegclassification_list}
\title{Plotclassification}
\usage{
\method{plot}{inbovegclassification_list}(x, ...)
}
\arguments{
\item{x}{list of classifications}

\item{...}{arguments passed to plot.inbovegclassification
and plot_classification_rose}
}
\value{
plot
}
\description{
Plotclassification
}
