% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{print.inbovegclassification}
\alias{print.inbovegclassification}
\title{Print method for inbovegclassification}
\usage{
\method{print}{inbovegclassification}(x, ...)
}
\arguments{
\item{x}{inbovegclassification object}

\item{...}{arguments passed to cat}
}
\description{
Print method for inbovegclassification
}
