% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{print.inbovegclassification_list}
\alias{print.inbovegclassification_list}
\title{Print method for inbovegclassification_list}
\usage{
\method{print}{inbovegclassification_list}(x, ...)
}
\arguments{
\item{x}{inbovegclassification_list object}

\item{...}{arguments passed to cat}
}
\description{
Print method for inbovegclassification_list
}
