% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{print.summary.inbovegclassification_list}
\alias{print.summary.inbovegclassification_list}
\title{Print method for summary.inbovegclassification_list}
\usage{
\method{print}{summary.inbovegclassification_list}(x, ...)
}
\arguments{
\item{x}{inbovegclassification_list object}

\item{...}{not used}
}
\description{
Print method for summary.inbovegclassification_list
}
