% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inboveg_recording.R
\name{read_inboveg_recording}
\alias{read_inboveg_recording}
\title{Read a recording from inboveg}
\usage{
read_inboveg_recording(con_inboveg, con_taxa, survey, code)
}
\arguments{
\item{con_inboveg}{db connection to Cydonia}

\item{con_taxa}{db connection to "Taxa" db}

\item{survey}{survey name}

\item{code}{releve code within survey}
}
\value{
data.frame
}
\description{
Read a recording from Cydonia, specifying the survey and RecordingGivid
}
\examples{
\dontrun{
library(inbovegtypering)
con_inboveg <- connect_db_inboveg()
con_taxa <- connect_db_taxonomy()
record <-
  read_inboveg_recording(con_inboveg, con_taxa,
    survey = "MILKLIM_Heischraal2012",
    code = "IV2012081611384756"
  )
}
}
