% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_getters.R
\name{read_inboveg_specieslist}
\alias{read_inboveg_specieslist}
\title{Read a species list from the taxonomy database}
\usage{
read_inboveg_specieslist(con_taxa)
}
\arguments{
\item{con_taxa}{db connection to "Taxa" db.}
}
\value{
A tibble (from dbplyr) with the species list.
}
\description{
Retrieves the species list (TaxonSourceTaxonGbifMatch) from the
taxonomy database.
}
