% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods_summary.R
\name{summary.inbovegclassification}
\alias{summary.inbovegclassification}
\title{Summary Method for inbovegclassification Objects}
\usage{
\method{summary}{inbovegclassification}(object, n = 5, ...)
}
\arguments{
\item{object}{An object of class "inbovegclassification"}

\item{n}{Number of top matches to show (default 5)}

\item{...}{Additional arguments (not used)}
}
\value{
A list object of class \code{summary.inbovegclassification}
}
\description{
Provides a summary of vegetation classification results,
including best matches and additional statistics.
}
