% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{assert_correct_synoptics}
\alias{assert_correct_synoptics}
\title{Check correct synoptics format}
\usage{
assert_correct_synoptics(x)
}
\arguments{
\item{x}{A data.frame to be assessed as a synoptic table.}
}
\value{
Throws an error if format is incorrect, otherwise returns \code{TRUE}.
}
\description{
Check correct synoptics format
}
\keyword{internal}
