% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{assert_db_recording_format}
\alias{assert_db_recording_format}
\title{Assert Database Recording Format}
\usage{
assert_db_recording_format(data)
}
\arguments{
\item{data}{A tibble or data frame containing the database recording data.}
}
\value{
Returns TRUE if the data matches the expected format.
Otherwise, it throws an error.
}
\description{
This function checks if the provided data matches the expected format
for database recordings.
}
\keyword{internal}
