% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_cod.R
\name{classify_cod}
\alias{classify_cod}
\title{Classify based on the Composite Distance (CoD)}
\usage{
classify_cod(data, synoptics, r = 1.5, s = 0.6, ...)
}
\arguments{
\item{data}{A dataframe of plot data.}

\item{synoptics}{A dataframe of synoptic data.}

\item{r}{Weight for Weirdness component (default 1.5).}

\item{s}{Power for MED component (default 0.60).}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class 'inbovegclassification_list'.
}
\description{
Implements Equation 9 from van Tongeren et al. (2008) \link[=cite: 1]{cite_start}.
Combines Weirdness, Incompleteness, and MED into a single index.
}
