% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_euclidean_old.R, R/classify_med.R
\name{classify_euclidean}
\alias{classify_euclidean}
\title{Calculate Modified Euclidean Distance}
\usage{
classify_euclidean(data, synoptics, t = 1, ...)

classify_euclidean(data, synoptics, t = 1, ...)
}
\arguments{
\item{data}{A dataframe of plot data.}

\item{synoptics}{A dataframe of synoptic data.}

\item{t}{Weight for absent species in MED (default 1).}

\item{...}{Additional arguments (not used).}
}
\value{
An inbovegclassification object

An object of class 'inbovegclassification_list'.
}
\description{
Calculates modified Euclidean distance following van Tongeren et al. (2008)

Implements Equation 8 from van Tongeren et al. (2008) \link[=cite: 1]{cite_start}.
Calculates a quantitative distance based on species abundance.
}
