% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_euclidean.R
\name{classify_euclidean}
\alias{classify_euclidean}
\title{Calculate Modified Euclidean Distance}
\usage{
classify_euclidean(data, synoptics, t = 0.5)
}
\arguments{
\item{data}{A dataframe containing the plot data with columns:
\itemize{
\item RecordingGivid: Unique identifier for the recording
\item LayerCode: Code identifying vegetation layer
\item CoverageCode: Code for coverage type
\item PctValue: Percentage cover value
\item gbif_usageKey: GBIF species identifier
}
on wich recordings are identified.
When NULL all data is assumed from same recording}

\item{synoptics}{A dataframe containing reference syntaxa data with columns:
\itemize{
\item syntaxoncode: Code identifying the vegetation type
\item usageKey: GBIF species identifier
\item frequentie: Frequency of species in vegetation type (0-1)
}
@param normalised ignored logical variable}

\item{t}{Weight for absent species (0-1, default 0.5)}
}
\value{
An inbovegclassification object
}
\description{
Calculates modified Euclidean distance following van Tongeren et al. (2008)
}
