% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_incompleteness.R
\name{classify_incompleteness}
\alias{classify_incompleteness}
\title{Calculate Incompleteness Index}
\usage{
classify_incompleteness(data, synoptics)
}
\arguments{
\item{data}{A dataframe containing the plot data with columns:
\itemize{
\item RecordingGivid: Unique identifier for the recording
\item LayerCode: Code identifying vegetation layer
\item CoverageCode: Code for coverage type
\item PctValue: Percentage cover value
\item gbif_usageKey: GBIF species identifier
}}

\item{synoptics}{A dataframe containing reference syntaxa data with columns:
\itemize{
\item syntaxoncode: Code identifying the vegetation type
\item usageKey: GBIF species identifier
\item frequentie: Frequency of species in vegetation type (0-1)
}}
}
\value{
An inbovegclassification object
}
\description{
Calculates incompleteness index following van Tongeren et al. (2008)
}
