% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_indices.R
\name{classify_likelihood}
\alias{classify_likelihood}
\title{Calculate -2ln(Likelihood) Index (Eq. 4)}
\usage{
classify_likelihood(analysis_df, ...)
}
\arguments{
\item{analysis_df}{A data frame prepared by \code{create_analysis_data()}.}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class \code{inbovegclassification_list}.
}
\description{
Internal function to calculate the likelihood index.
This is the sum of the Weirdness and Incompleteness indices.
}
\keyword{internal}
