% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_weirdness.R,
%   R/classify_weirdness_old.R
\name{classify_weirdness}
\alias{classify_weirdness}
\title{Classify based on Weirdness Index}
\usage{
classify_weirdness(data, synoptics = "default", normalised = TRUE)

classify_weirdness(data, synoptics = "default", normalised = TRUE)
}
\arguments{
\item{data}{A dataframe of plot data.}

\item{synoptics}{A dataframe of synoptic data.}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class 'inbovegclassification_list'.

An inbovegclassification object
}
\description{
Implements Equation 5 from van Tongeren et al. (2008) \link[=cite: 1]{cite_start}.
Calculates the sum of contributions from species present in the relevé.

Calculates weirdness index following van Tongeren et al. (2008)
}
