% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_indices.R
\name{classify_weirdness}
\alias{classify_weirdness}
\title{Calculate Weirdness Index (Eq. 5)}
\usage{
classify_weirdness(analysis_df, ...)
}
\arguments{
\item{analysis_df}{A data frame prepared by \code{create_analysis_data()}.}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class \code{inbovegclassification_list}.
}
\description{
Internal function to calculate the Weirdness index.
Calculates the sum of contributions from species \emph{present} in the relevé.
}
\keyword{internal}
