% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db_connect.R
\name{connect_db_flora}
\alias{connect_db_flora}
\title{Connect to INBO flora database}
\usage{
connect_db_flora(db = "D0012_00_Flora", test = FALSE)
}
\arguments{
\item{db}{Character string specifying the database name.
Defaults to "D0012_00_Flora".}

\item{test}{flag whether the test modus is used}
}
\value{
A database connection object
}
\description{
Creates a database connection to the INBO flora database.
}
