% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_connections.R
\name{connect_db_inboveg}
\alias{connect_db_inboveg}
\title{Connect to the INBO database}
\usage{
connect_db_inboveg(db = "D0010_00_Cydonia", test = FALSE)
}
\arguments{
\item{db}{The name of the INBO hosted database. Defaults to Cydonia db}

\item{test}{flag whether the test modus is used}
}
\value{
connection object
}
\description{
Connect to the INBO database
}
