% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db_connect.R
\name{connect_db_taxonomy}
\alias{connect_db_taxonomy}
\title{Connect to INBO taxonomy database}
\usage{
connect_db_taxonomy(db = "D0155_00_Taxa", test = FALSE)
}
\arguments{
\item{db}{Character string specifying the database name.
Defaults to "D0155_00_Taxa".}

\item{test}{Flag whether the test modus is used. If TRUE, tries to connect
to a local test database.}
}
\value{
A database connection object.
}
\description{
Creates a database connection to the INBO taxonomy database.
}
