% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_recordings}
\alias{example_recordings}
\title{Example vegetation recordings}
\format{
A data frame with example variables:
\describe{
\item{Name}{Name of the project}
\item{RecordingGivid}{Name of the record}
\item{UserReference}{Naming in the project}
\item{LayerCode}{Layer identifier}
\item{CoverCode}{Coverage percentage of the layer}
\item{OriginalName}{Original scientific name}
\item{ScientificName}{Correct scientific name}
\item{TaxonGroupCode}{Grouping of species}
\item{PctValue}{Percentage coverage over the whole plot}
\item{RecordingScale}{Scale on which the recordings are written down}
\item{species_number}{List column with matching synoptic species numbers}
... and other columns from inbodb.
}
}
\source{
Generated sample data from Cydonia database in INBO.
}
\usage{
example_recordings
}
\description{
A dataset containing sample vegetation recordings for testing and examples.
The data is retrieved from the INBOVEG database Cydonia.
}
\keyword{datasets}
