% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_synoptics}
\alias{example_synoptics}
\title{Example synoptic classification table}
\format{
A data frame with 4 variables:
\describe{
\item{syntaxonCode}{Syntaxon identifier}
\item{speciesNumber}{Species identifier (links to species list)}
\item{frequency}{Presence frequency of a species (0-100)}
\item{mean_if_present}{The average cover when present (0-100)}
}
}
\source{
Subset of syntaxons for classification.
}
\usage{
example_synoptics
}
\description{
A dataset containing the necessary info for classification of vegetation
in syntaxons.
}
\keyword{datasets}
