% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_recording_species_list.R
\name{link_recording_species_list}
\alias{link_recording_species_list}
\title{Link recording to species list}
\usage{
link_recording_species_list(
  data,
  col_name,
  con = NULL,
  source = "taxon_db",
  taxons = NULL,
  max_gbif = 5000
)
}
\arguments{
\item{data}{A dataframe containing recordings}

\item{col_name}{Character string specifying the column name in 'data'
that contains the species names to match}

\item{con}{A database connection object to the INBO taxonomy database}

\item{source}{where to find the taxonomic information:
"included" the usageKey and acceptedUsageKey are already present.
Used to check if the data is compatible for classification
"taxon_db" INBO taxon db,
"gbif_backbone" get from via gbif backbone api,
"data.frame" data.frame containing the variables}

\item{taxons}{the data.frame containing name, usageKey, acceptedUsageKey
semicolon separated, with these names: name, usageKey, acceptedUsageKey}

\item{max_gbif}{integer >= 0 indicating the max species provided to gbif,
so not accidentally thousands of records are requested by api. Default 5000.}
}
\value{
A dataframe containing the original data
with an additional column
'gbif_usageKey' that contains the GBIF taxonomy keys for matched species
}
\description{
This function links species records to their corresponding
GBIF taxonomy information needed in this package
}
\examples{
\dontrun{
# Assuming 'con' is your database connection
# and 'species_data' is your dataframe with a column 'species_name'
library(inbovegtypering)
con <- connect_db_taxonomy()
link_recording_species_list(
  example_recordings |> select(2, 4, 6, 12),
  "OriginalName",
  con = con,
  source = "taxon_db")

test_records <- example_recordings |> slice(1:5) |> select(2, 4, 6, 12)
link_recording_species_list(test_records, col_name = "OriginalName",
                            source = "gbif_backbone")
}

library(inbovegtypering)
test_records <- example_recordings |> slice(c(1:3,5)) |> select(2, 4, 6, 12)
species_data <- data.frame(name = c("Acer platanoides L.",
                                    "Acer pseudoplatanus L.",
                                    "Achillea millefolium L.",
                                    "Achillea ptarmica L."),
                           usageKey = c(3189846, 3189870, 3120060, 3120333),
                           acceptedUsageKey = NA)
taxons <-
  lapply(species_data$usageKey,
         FUN = function(x) {rgbif::name_usage(x)$data}) |>
  bind_rows() |>
  right_join(species_data, join_by(key == usageKey)) |>
  mutate(usageKey = key,
         acceptedUsageKey = NA_character_)
link_recording_species_list(test_records, col_name = "OriginalName",
                            source = "data.frame", taxons = taxons)

}
\seealso{
\code{\link{connect_db_taxonomy}} for establishing the database connection
}
