% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_linking.R
\name{link_species_list}
\alias{link_species_list}
\title{Link species with internal species list}
\usage{
link_species_list(species_names)
}
\arguments{
\item{species_names}{character vector of species names to be found}
}
\value{
data.frame containing species information
}
\description{
Link species with internal species list
}
\keyword{internal}
