% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_linking.R
\name{link_synoptic_species}
\alias{link_synoptic_species}
\title{Find corresponding species_number in the synoptic species list}
\usage{
link_synoptic_species(usageKey, acceptedUsageKey, name)
}
\arguments{
\item{usageKey}{The species' gbif_usageKey.}

\item{acceptedUsageKey}{The species' gbif_acceptedusageKey.}

\item{name}{The species' original name (as fallback).}
}
\value{
A vector of matching \code{speciesNumber} values, or \code{NA_integer_}
if no match.
}
\description{
Find corresponding species_number in the synoptic species list
}
\keyword{internal}
