% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_linking.R
\name{link_taxa_db}
\alias{link_taxa_db}
\title{Link species names to INBO taxonomy database}
\usage{
link_taxa_db(con, species_names)
}
\arguments{
\item{con}{A database connection object to the INBO taxonomy database.}

\item{species_names}{A character vector containing species records.}
}
\value{
A dataframe containing the matched species from the futondb.
}
\description{
Link species names to INBO taxonomy database
}
\keyword{internal}
