% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inbovegclassification.R
\name{plot.inbovegclassification}
\alias{plot.inbovegclassification}
\title{Plot Method for inbovegclassification Objects}
\usage{
\method{plot}{inbovegclassification}(x, n_types = 12, ...)
}
\arguments{
\item{x}{An object of class "inbovegclassification"}

\item{n_types}{Number of types to show (default 12)}

\item{...}{Additional arguments passed to methods}
}
\value{
A ggplot object
}
\description{
Generic plot method that dispatches to appropriate plotting function
based on indextype
}
\examples{
\dontrun{
results <- classify_likelihood(data, synoptics)
plot(results)
plot(results, index = "likelihood") # likelihood plot
}
}
