% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods_plot.R
\name{plot.inbovegclassification_list}
\alias{plot.inbovegclassification_list}
\title{Plot Method for inbovegclassification_list Objects}
\usage{
\method{plot}{inbovegclassification_list}(x, ...)
}
\arguments{
\item{x}{list of classifications}

\item{...}{arguments passed to \code{plot.inbovegclassification}
and \code{plot_classification_rose}}
}
\value{
This function prints plots but returns \code{invisible(NULL)}.
}
\description{
Plots the classification rose for each relevé in the list.
In a non-interactive session, this will print plots one after another.
In an interactive session (like RStudio), it will show them sequentially.
}
