% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification_rose.R
\name{plot_classification_rose}
\alias{plot_classification_rose}
\title{Create Classification Rose Plot}
\usage{
plot_classification_rose(
  data,
  indextype,
  types = 12,
  min_length = 0.5,
  order = c("likelihood", "alphabetical", "given"),
  ...
)
}
\arguments{
\item{data}{A tibble with classification results}

\item{indextype}{Character string indicating the type of index}

\item{types}{Either a number indicating how many top types to show,
or a character vector of syntaxon codes to display}

\item{min_length}{Minimum length for normalization (default 0.5)}

\item{order}{on which should the classes be ordered, on the best match,
alphabetical or as given in the order of the types argument}

\item{...}{Additional arguments passed to methods}
}
\value{
A ggplot object
}
\description{
Creates a star-shaped radial plot showing classification results
}
\examples{
\dontrun{
plot(results, n_types = 8)
selected_types <- c("type1", "type2", "type3")
plot(results, types = selected_types)
library(patchwork)
record1_plot <- plot(results1, types = selected_types)
record2_plot <- plot(results2, types = selected_types)
record1_plot + record2_plot
}

}
\keyword{internal}
