% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods_print.R
\name{print.inbovegclassification}
\alias{print.inbovegclassification}
\title{Print Method for inbovegclassification Objects}
\usage{
\method{print}{inbovegclassification}(x, n = 10, ...)
}
\arguments{
\item{x}{An object of class "inbovegclassification"}

\item{n}{Number of top matches to show (default 10)}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Prints the results of vegetation classification, showing the top matches.
}
