% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods_print.R
\name{print.inbovegclassification_list}
\alias{print.inbovegclassification_list}
\title{Print method for inbovegclassification_list Objects}
\usage{
\method{print}{inbovegclassification_list}(x, ...)
}
\arguments{
\item{x}{list of classifications}

\item{...}{arguments passed to print.inbovegclassification}
}
\value{
Invisibly returns the input object
}
\description{
Prints the classification results for each relevé in the list.
}
