% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_inbovegclassification.R
\name{print.inbovegclassification_list}
\alias{print.inbovegclassification_list}
\title{Print classification}
\usage{
\method{print}{inbovegclassification_list}(x, ...)
}
\arguments{
\item{x}{list of classifications}

\item{...}{arguments passed to print.inbovegclassification}
}
\value{
print output
}
\description{
Print classification
}
