% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_inbovegclassification.R
\name{print.summary.inbovegclassification}
\alias{print.summary.inbovegclassification}
\title{Print Method for summary.inbovegclassification Objects}
\usage{
\method{print}{summary.inbovegclassification}(x, ...)
}
\arguments{
\item{x}{A summary.inbovegclassification object}

\item{...}{Additional arguments passed to methods}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for summary.inbovegclassification Objects
}
