% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{print.summary.inbovegclassification}
\alias{print.summary.inbovegclassification}
\title{Print method for summary.inbovegclassification}
\usage{
\method{print}{summary.inbovegclassification}(x, ...)
}
\arguments{
\item{x}{inbovegclassification object}

\item{...}{not used}
}
\description{
Print method for summary.inbovegclassification
}
