% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_getters.R
\name{read_inboveg_recording}
\alias{read_inboveg_recording}
\title{Read a recording from inboveg}
\usage{
read_inboveg_recording(con_inboveg, con_taxa, survey, code)
}
\arguments{
\item{con_inboveg}{db connection to Cydonia (inboveg).}

\item{con_taxa}{db connection to "Taxa" db.}

\item{survey}{survey name.}

\item{code}{releve code (RecordingGivid) within survey.}
}
\value{
A data.frame with the relevé data, linked to taxonomic info
(including \code{species_number}).
}
\description{
Read a recording from Cydonia, specifying the survey and RecordingGivid,
and link it to taxonomic information.
}
