% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_inbovegclassification.R
\name{summary.inbovegclassification}
\alias{summary.inbovegclassification}
\title{Summary Method for inbovegclassification Objects}
\usage{
\method{summary}{inbovegclassification}(
  object,
  n_types = 10,
  types = NULL,
  order = c("likelihood", "alphabetical", "given"),
  ...
)
}
\arguments{
\item{object}{An object of class "inbovegclassification"}

\item{n_types}{Number of types to show (default 10)}

\item{types}{Optional character vector of specific syntaxon codes to show}

\item{order}{Ordering of results:
"likelihood" (default), "alphabetical", or "given"}

\item{...}{Additional arguments passed to methods}
}
\value{
A list with summary statistics
}
\description{
Provides a summary of vegetation classification results,
including best matches and additional statistics.
}
