% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods_summary.R
\name{summary.inbovegclassification_list}
\alias{summary.inbovegclassification_list}
\title{Summary method for inbovegclassification_list Objects}
\usage{
\method{summary}{inbovegclassification_list}(object, ...)
}
\arguments{
\item{object}{list of classifications}

\item{...}{arguments passed to summary.inbovegclassification}
}
\value{
A list of summary objects
}
\description{
Applies \code{summary()} to each classification in the list.
}
