% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_inbovegclassification.R
\name{summary.inbovegclassification_list}
\alias{summary.inbovegclassification_list}
\title{Summarise classification}
\usage{
\method{summary}{inbovegclassification_list}(x, ...)
}
\arguments{
\item{x}{list of classifications}

\item{...}{arguments passed to summary.inbovegclassification}
}
\value{
summary output
}
\description{
Summarise classification
}
